Create search bar section and display search bar
A div was created for the upper half of the header to allow for future
formatting. A div was also created for the search bar input and label
to allow for future formatting. An icon directory was created to house
the search icon. The search bar and search icon were displayed to the
webpage.